# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.ErrorResponse:
    properties:
      error: {}
    type: object
...
  controllers.SimpleMessageResponse:
    properties:
      response: {}
      status:
        type: integer
    type: object
...
  controllers.UserCreateRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  controllers.UserResponse:
    properties:
      avatar_path:
        type: string
      email:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      username:
        type: string
    type: object
  controllers.UserResponseEnvelope:
    properties:
      response:
        $ref: '#/definitions/controllers.UserResponse'
      status:
        type: integer
    type: object
  controllers.UserUpdateRequest:
    properties:
      current_password:
        type: string
      email:
        type: string
      new_password:
        type: string
    type: object
  controllers.UsersListResponse:
    properties:
      response:
        items:
          $ref: '#/definitions/controllers.UserResponse'
        type: array
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: API for brackets, matchups, voting, and engagement
  title: Matchup API
  version: "1.0"
paths:
...
  /me:
    get:
      description: Get the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UserResponseEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
...
  /users:
    get:
      description: Get a list of users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UsersListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.UserResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Register a user
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete the authenticated user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SimpleMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Get a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UserResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update email or password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UserResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
  /users/{id}/avatar:
    put:
      consumes:
      - multipart/form-data
      description: Upload a new avatar image
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Avatar image
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.UserResponseEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update avatar
      tags:
      - users
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Provide a valid JWT as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"